Add support for QR Analytics
closes #743
