+++ apps/web/app/api/links/iframeable/route.ts
@@ -1,5 +1,5 @@
 import { DubApiError, handleAndReturnErrorResponse } from "@/lib/api/errors";
-import { getIdentityHash } from "@/lib/edge";
+import { getIdentityHash } from "@/lib/middleware/utils";
 import { ratelimit } from "@/lib/upstash";
 import { getDomainQuerySchema, getUrlQuerySchema } from "@/lib/zod/schemas";
 import { isIframeable } from "@dub/utils";
