+++ apps/web/app/api/metatags/route.ts
@@ -1,5 +1,5 @@
 import { DubApiError, handleAndReturnErrorResponse } from "@/lib/api/errors";
-import { getIdentityHash } from "@/lib/edge";
+import { getIdentityHash } from "@/lib/middleware/utils";
 import { ratelimit } from "@/lib/upstash";
 import { getUrlQuerySchema } from "@/lib/zod/schemas";
 import { getToken } from "next-auth/jwt";
