+++ apps/web/app/api/providers/route.ts
@@ -1,5 +1,5 @@
 import { handleAndReturnErrorResponse } from "@/lib/api/errors";
-import { getIdentityHash } from "@/lib/edge";
+import { getIdentityHash } from "@/lib/middleware/utils";
 import { ratelimit } from "@/lib/upstash";
 import { getUrlQuerySchema } from "@/lib/zod/schemas";
 import { fetchWithTimeout } from "@dub/utils";
