+++ apps/web/lib/middleware/utils/detect-bot.ts
@@ -1,0 +1,20 @@
+import { NextRequest } from "next/server";
+
+export const detectBot = (req: NextRequest) => {
+  const url = req.nextUrl;
+  if (url.searchParams.get("bot")) return true;
+  const ua = req.headers.get("User-Agent");
+  if (ua) {
+    /* Note:
+     * - bot is for most bots & crawlers
+     * - ChatGPT is for ChatGPT
+     * - facebookexternalhit is for Facebook crawler
+     * - WhatsApp is for WhatsApp crawler
+     * - MetaInspector is for https://metatags.io/
+     */
+    return /bot|chatgpt|facebookexternalhit|WhatsApp|google|baidu|bing|msn|duckduckbot|teoma|slurp|yandex|MetaInspector/i.test(
+      ua,
+    );
+  }
+  return false;
+};
