+++ apps/web/lib/middleware/utils/detect-qr.ts
@@ -1,0 +1,7 @@
+import { NextRequest } from "next/server";
+
+export const detectQr = (req: NextRequest) => {
+  const url = req.nextUrl;
+  if (url.searchParams.get("qr") === "1") return true;
+  return false;
+};
