+++ apps/web/lib/middleware/utils/get-final-url.ts
@@ -1,0 +1,28 @@
+import { NextRequest } from "next/server";
+
+export const getFinalUrl = (url: string, { req }: { req: NextRequest }) => {
+  // query is the query string (e.g. d.to/github?utm_source=twitter -> ?utm_source=twitter)
+  const searchParams = req.nextUrl.searchParams;
+
+  // get the query params of the target url
+  const urlObj = new URL(url);
+
+  // if there are no query params, then return the target url as is (no need to parse it)
+  // @ts-ignore – until https://github.com/microsoft/TypeScript/issues/54466 is fixed
+  if (searchParams.size === 0) return url;
+
+  // if searchParams (type: `URLSearchParams`) has the same key as target url, then overwrite it
+  for (const [key, value] of searchParams) {
+    urlObj.searchParams.set(key, value);
+  }
+
+  if (urlObj.searchParams.get("qr") === "1") {
+    // remove qr param from the final url (only used for detectQr function)
+    urlObj.searchParams.delete("qr");
+  }
+
+  // construct final url
+  const finalUrl = urlObj.toString();
+
+  return finalUrl;
+};
