+++ apps/web/lib/middleware/utils/get-identity-hash.ts
