+++ apps/web/lib/planetscale.ts
@@ -1,4 +1,4 @@
-import { nanoid } from "@dub/utils";
+import { nanoid, punyEncode } from "@dub/utils";
 import { connect } from "@planetscale/database";
 import { DomainProps, WorkspaceProps } from "./types";
 
@@ -53,7 +53,7 @@ export const getLinkViaEdge = async (domain: string, key: string) => {
   const { rows } =
     (await conn.execute(
       "SELECT * FROM Link WHERE domain = ? AND `key` = ?",
-      [domain, decodeURIComponent(key)], // we need to make sure that the key is always decoded (cause that's how we store it in MySQL)
+      [domain, punyEncode(decodeURIComponent(key))], // we need to make sure that the key is always URI-decoded + punycode-encoded (cause that's how we store it in MySQL)
     )) || {};
 
   return rows && Array.isArray(rows) && rows.length > 0
