+++ packages/utils/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@dub/utils",
   "description": "Utility functions and constants for Dub.co",
-  "version": "0.0.47",
+  "version": "0.0.48",
   "sideEffects": false,
   "main": "./dist/index.mjs",
   "module": "./dist/index.mjs",
